Avoided some warnings with GCC -Wconversion and enabled -Wconversion in configure.ac. Fuego should now compile without warnings on 64-bit systems with SgUctValue=double.
git-svn-id: svn://svn.code.sf.net/p/fuego/code/trunk@1224 2e953b5c-c64d-0410-be54-f773e93e544c

+++ smartgame/SgThreadedWorker.h
@@ -87,8 +87,8 @@ private:
 template<typename I, typename O, typename W>
 SgThreadedWorker<I,O,W>::SgThreadedWorker(std::vector<W>& workers)
     : m_quit(false),
-      m_startWork(workers.size() + 1),
-      m_workFinished(workers.size() + 1)
+      m_startWork(static_cast<unsigned int>(workers.size() + 1)),
+      m_workFinished(static_cast<unsigned int>(workers.size() + 1))
 {
     for (std::size_t i = 0; i < workers.size(); ++i)
     {
